Move article editor into apps directory, reorganize repo structure

+++ server/src/main/java/wiki/ednotes/server/article/ArticleRepository.java
@@ -3,6 +3,6 @@ package wiki.ednotes.server.article;
 import org.springframework.data.jpa.repository.JpaRepository;
 import java.util.List;
 
-public interface ArticleRepository extends JpaRepository<Article, Integer> {
-   List<Article> findByCategoryIdOrderByOrderAsc(Integer categoryId);
+public interface ArticleRepository extends JpaRepository<Article, Long> {
+   List<Article> findByCategoryIdOrderByOrderAsc(Long categoryId);
 }
