+++ server/src/main/java/wiki/ednotes/server/category/CategoryRepository.java
@@ -6,7 +6,7 @@ import java.util.List;
 /**
  * Repository interface for managing categories.
  */
-public interface CategoryRepository extends JpaRepository<Category, Integer> {
-    List<Category> findByParentIdOrderByOrderAsc(Integer parentId);
-    List<Category> findByParentIdAndPublishedIsTrueOrderByOrderAsc(Integer parentId);
+public interface CategoryRepository extends JpaRepository<Category, Long> {
+    List<Category> findByParentIdOrderByOrderAsc(Long parentId);
+    List<Category> findByParentIdAndPublishedIsTrueOrderByOrderAsc(Long parentId);
 }
